Allow tracking of larger contribution amounts by increasing data type
Update the 'amount' column in the Contribution model from Integer to BigInteger to accommodate larger numerical values, resolving the NumericValueOutOfRange error.

+++ models.py
@@ -72,7 +72,7 @@ class Contribution(Base):
     guild_id = Column(BigInteger, ForeignKey("guilds.id"), nullable=False)
     member_id = Column(BigInteger, ForeignKey("members.id"), nullable=False)
     material_id = Column(Integer, ForeignKey("materials.id"), nullable=False)
-    amount = Column(Integer, nullable=False)
+    amount = Column(BigInteger, nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow)
 
     # Relationships
